p1073: move chart-container, chart-with-loader to subfolders
