+++ src/domains/chart/components/chart-with-loader/chart-with-loader.css
