+++ src/domains/chart/components/chart-with-loader/index.ts
@@ -1,0 +1,2 @@
+export * from "./chart-with-loader"
+
